# api/openapi.yaml
openapi: 3.1.0
info:
  title: Domestique API
  version: 2.0.0
  description: |
    The API contract served by Domestique. JSON property names are camelCase.
    All routes except `GET /healthz` require the browser session cookie;
    state-changing routes also require the same-origin `Origin` header. The
    sign-in routes under `/auth/` are browser navigation rather than API
    operations, and are not described here.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /v1/status:
    get:
      tags:
        - browser-api
      operationId: getStatus
      responses:
        "200":
          description: Current local status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/tasks:
    get:
      tags:
        - browser-api
      operationId: listTasks
      security:
        - session: []
      description: >-
        Every background activity this build registers, in the order a status
        page reads best.
      responses:
        "200":
          description: The registered tasks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /v1/tasks/{name}/schedule:
    put:
      tags:
        - browser-api
      operationId: setTaskSchedule
      security:
        - session: []
          browserOrigin: []
      description: >-
        Sets whether the schedule may start this task. It governs unattended
        runs only: a task switched off still runs when an operator asks for it.
      parameters:
        - $ref: '#/components/parameters/TaskName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskScheduleUpdate'
      responses:
        "200":
          description: The registered tasks, as they now stand.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/tasks/{name}/run:
    post:
      tags:
        - browser-api
      operationId: runTask
      security:
        - session: []
          browserOrigin: []
      description: >-
        Starts one attempt of a task and reports whether it was accepted. An
        accepted attempt continues independently of this request.
      parameters:
        - $ref: '#/components/parameters/TaskName'
      responses:
        "202":
          $ref: '#/components/responses/Accepted'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/TaskInProgress'
  /v1/tasks/{name}/run/{argument}:
    post:
      tags:
        - browser-api
      operationId: runTaskArgument
      security:
        - session: []
          browserOrigin: []
      description: >-
        Starts one attempt of a task over one argument, such as a single target
        slot. Otherwise identical to running it with none. The `404` is about
        the task name; an unconfigured argument is the task's own to interpret
        and is accepted.
      parameters:
        - $ref: '#/components/parameters/TaskName'
        - $ref: '#/components/parameters/TaskArgument'
      responses:
        "202":
          $ref: '#/components/responses/Accepted'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/TaskInProgress'
  /v1/tasks/runs:
    get:
      tags:
        - browser-api
      operationId: getTaskRuns
      description: >-
        What the background activities have been doing, newest first: one page
        of the recorded history, over every task or over one named task.
      parameters:
        - name: task
          in: query
          description: >-
            Narrows the page to one registered task. A name this build does not
            register is refused rather than answered with an empty page.
          schema:
            type: string
        - name: after
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A newest-first cursor page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunPage'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/sync/runs:
    get:
      tags:
        - browser-api
      operationId: getSyncRuns
      parameters:
        - name: after
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A newest-first cursor page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRunPage'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/routes:
    get:
      tags:
        - browser-api
      operationId: getRoutes
      responses:
        "200":
          description: Stored route inventory without geometry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/providers/{provider}/sourceRoutes/{sourceRouteId}/routes/{stageOrder}:
    get:
      tags:
        - browser-api
      operationId: getRoute
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/SourceRouteId'
        - $ref: '#/components/parameters/StageOrder'
      responses:
        "200":
          description: Stored route metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/providers/{provider}/sourceRoutes/{sourceRouteId}/routes/{stageOrder}/geometry:
    get:
      tags:
        - browser-api
      operationId: getRouteGeometry
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/SourceRouteId'
        - $ref: '#/components/parameters/StageOrder'
      responses:
        "200":
          description: Stored route geometry.
          content:
            application/geo+json:
              schema:
                $ref: '#/components/schemas/GeoJSONFeature'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/providers/{provider}/sourceRoutes/{sourceRouteId}/routes/{stageOrder}/reprocess:
    post:
      tags:
        - browser-api
      operationId: reprocessRoute
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/SourceRouteId'
        - $ref: '#/components/parameters/StageOrder'
      security:
        - session: []
          browserOrigin: []
      responses:
        "202":
          $ref: '#/components/responses/Accepted'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "503":
          $ref: '#/components/responses/Unavailable'
  /v1/providers/{provider}/routes/{routeId}/stages/{stage}:
    get:
      operationId: redirectStageRoute
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/LegacyRouteId'
        - $ref: '#/components/parameters/LegacyStage'
      responses:
        "308":
          $ref: '#/components/responses/LegacyRedirect'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/providers/{provider}/routes/{routeId}/stages/{stage}/geometry:
    get:
      operationId: redirectStageGeometry
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/LegacyRouteId'
        - $ref: '#/components/parameters/LegacyStage'
      responses:
        "308":
          $ref: '#/components/responses/LegacyRedirect'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/providers/{provider}/routes/{routeId}/stages/{stage}/reprocess:
    post:
      operationId: redirectStageReprocess
      parameters:
        - $ref: '#/components/parameters/Provider'
        - $ref: '#/components/parameters/LegacyRouteId'
        - $ref: '#/components/parameters/LegacyStage'
      security:
        - session: []
          browserOrigin: []
      responses:
        "308":
          $ref: '#/components/responses/LegacyRedirect'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/routes/{routeId}/stages/{stage}:
    get:
      operationId: redirectLegacyRoute
      parameters:
        - $ref: '#/components/parameters/LegacyRouteId'
        - $ref: '#/components/parameters/LegacyStage'
      responses:
        "308":
          $ref: '#/components/responses/LegacyRedirect'
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-domestique_session
      description: Browser session cookie issued by the sign-in flow.
    browserOrigin:
      type: apiKey
      in: header
      name: Origin
      description: Exact browser origin configured for this service. Required by every operation that starts a run or writes state, and modelled as a security scheme rather than a parameter because a request without it is refused rather than misread.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid browser session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Status:
      type: object
      additionalProperties: false
      required:
        - ready
        - converged
        - targets
        - sync
      properties:
        ready:
          type: boolean
        converged:
          type: boolean
        build:
          $ref: '#/components/schemas/Build'
        targets:
          type: array
          items:
            $ref: '#/components/schemas/TargetStatus'
        sync:
          $ref: '#/components/schemas/SyncStatus'
...